Move declarations into utils file
